+++ models/lenet.py
@@ -21,18 +21,18 @@ class LeNet5(nn.Module):
 
         self.convnet = nn.Sequential(OrderedDict([
             ('c1', nn.Conv2d(1, 6, kernel_size=(5, 5))),
-            ('relu1', nn.Tanh()),
+            ('tanh1', nn.Tanh()),
             ('s2', nn.MaxPool2d(kernel_size=(2, 2), stride=2)),
             ('c3', nn.Conv2d(6, 16, kernel_size=(5, 5))),
-            ('relu3', nn.Tanh()),
+            ('tanh3', nn.Tanh()),
             ('s4', nn.MaxPool2d(kernel_size=(2, 2), stride=2)),
             ('c5', nn.Conv2d(16, 120, kernel_size=(5, 5))),
-            ('relu5', nn.Tanh())
+            ('tanh5', nn.Tanh())
         ]))
 
         self.fc = nn.Sequential(OrderedDict([
             ('f6', nn.Linear(120, 84)),
-            ('relu6', nn.Tanh()),
+            ('tanh6', nn.Tanh()),
             ('f7', nn.Linear(84, 10)),
             ('sig7', nn.LogSoftmax(dim=-1))
         ]))
